File for SSTScanSoftLimit
