+++ dbms/src/Storages/DeltaMerge/Decode/SSTFilesToBlockInputStream.h
@@ -17,6 +17,7 @@
 #include <DataStreams/IBlockInputStream.h>
 #include <RaftStoreProxyFFI/ColumnFamily.h>
 #include <Storages/DeltaMerge/DMVersionFilterBlockInputStream.h>
+#include <Storages/DeltaMerge/Decode/SSTScanSoftLimit.h>
 #include <Storages/KVStore/Decode/PartitionStreams.h>
 #include <Storages/KVStore/MultiRaft/PreHandlingTrace.h>
 
@@ -41,52 +42,6 @@ class SSTFilesToBlockInputStream;
 using SSTFilesToBlockInputStreamPtr = std::shared_ptr<SSTFilesToBlockInputStream>;
 class BoundedSSTFilesToBlockInputStream;
 using BoundedSSTFilesToBlockInputStreamPtr = std::shared_ptr<BoundedSSTFilesToBlockInputStream>;
-
-struct SSTScanSoftLimit
-{
-    constexpr static size_t HEAD_OR_ONLY_SPLIT = SIZE_MAX;
-    size_t split_id;
-    TiKVKey raw_start;
-    TiKVKey raw_end;
-    DecodedTiKVKey decoded_start;
-    DecodedTiKVKey decoded_end;
-    std::optional<RawTiDBPK> start_limit;
-    std::optional<RawTiDBPK> end_limit;
-
-    SSTScanSoftLimit(size_t split_id_, TiKVKey && raw_start_, TiKVKey && raw_end_)
-        : split_id(split_id_)
-        , raw_start(std::move(raw_start_))
-        , raw_end(std::move(raw_end_))
-    {
-        if (!raw_start.empty())
-        {
-            decoded_start = RecordKVFormat::decodeTiKVKey(raw_start);
-        }
-        if (!raw_end.empty())
-        {
-            decoded_end = RecordKVFormat::decodeTiKVKey(raw_end);
-        }
-        if (!decoded_start.empty())
-        {
-            start_limit = RecordKVFormat::getRawTiDBPK(decoded_start);
-        }
-        if (!decoded_end.empty())
-        {
-            end_limit = RecordKVFormat::getRawTiDBPK(decoded_end);
-        }
-    }
-
-    SSTScanSoftLimit clone() const { return SSTScanSoftLimit(split_id, raw_start.toString(), raw_end.toString()); }
-
-    const std::optional<RawTiDBPK> & getStartLimit() const { return start_limit; }
-
-    const std::optional<RawTiDBPK> & getEndLimit() const { return end_limit; }
-
-    std::string toDebugString() const
-    {
-        return fmt::format("{}:{}", raw_start.toDebugString(), raw_end.toDebugString());
-    }
-};
 
 struct SSTFilesToBlockInputStreamOpts
 {
